Minor refactor, fix demo app
